+++ packages/aptos-sdk/src/common/index.ts
@@ -8,7 +8,7 @@ export const APT_DIVIDER = toBn(10).pow(8)
 export const MAX_U64_INT = "18446744073709551615"
 
 export const APT_TOKEN_TYPE = "0x1::aptos_coin::AptosCoin"
-
+export const ENCODED_APT_TOKEN_TYPE = btoa(APT_TOKEN_TYPE)
 export type CURRENCY_TYPE = typeof APT_TOKEN_TYPE
 
 export function isChangeBelongsToType(change: WriteSetChange, dataTypeFn: (dataType: string) => boolean): boolean {